+++ sampleapp/src/main/java/io/matthewnelson/sampleapp/topl_android/MyEventBroadcaster.kt
@@ -69,8 +69,8 @@ package io.matthewnelson.sampleapp.topl_android
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import io.matthewnelson.sampleapp.ui.fragments.home.LogMessageAdapter
-import io.matthewnelson.topl_service.service.components.onionproxy.model.TorPortInfo
-import io.matthewnelson.topl_service.service.components.onionproxy.model.TorServiceEventBroadcaster
+import io.matthewnelson.topl_service_base.TorPortInfo
+import io.matthewnelson.topl_service_base.TorServiceEventBroadcaster
 import io.matthewnelson.topl_service_base.ServiceUtilities
 
 /**
