+++ topl-service/src/main/java/io/matthewnelson/topl_service/TorServiceController.kt
@@ -75,17 +75,13 @@ import io.matthewnelson.topl_core_base.TorConfigFiles
 import io.matthewnelson.topl_core_base.TorSettings
 import io.matthewnelson.topl_service.service.BaseService
 import io.matthewnelson.topl_service.lifecycle.BackgroundManager
-import io.matthewnelson.topl_service_base.TorServicePrefs
 import io.matthewnelson.topl_service.service.components.actions.ServiceActionProcessor
 import io.matthewnelson.topl_service.service.components.actions.ServiceAction
 import io.matthewnelson.topl_service.service.components.binding.TorServiceConnection
 import io.matthewnelson.topl_service.service.components.onionproxy.ServiceTorSettings
-import io.matthewnelson.topl_service.service.components.onionproxy.model.TorServiceEventBroadcaster
 import io.matthewnelson.topl_service.util.ServiceConsts
 import io.matthewnelson.topl_service.util.V3ClientAuthManager
-import io.matthewnelson.topl_service_base.ApplicationDefaultTorSettings
-import io.matthewnelson.topl_service_base.BaseServiceTorSettings
-import io.matthewnelson.topl_service_base.BaseV3ClientAuthManager
+import io.matthewnelson.topl_service_base.*
 
 class TorServiceController private constructor(): ServiceConsts() {
